refactor: rename power to arch
